+++ features/step_definitions/local-time-difference-steps.js
@@ -1,8 +1,11 @@
 import { Given, Then } from 'cucumber';
 
 Given(/^I set wrong local time difference$/, async function () {
-  await this.client.execute(timeDifference => {
-    daedalus.api.ada.setLocalTimeDifference(timeDifference);
+  await this.client.executeAsync((timeDifference, done) => {
+    daedalus.api.ada.setLocalTimeDifference(timeDifference)
+      .then(() => daedalus.stores.networkStatus._updateLocalTimeDifference())
+      .then(done)
+      .catch((error) => done(error));
   }, 1511823600000);
 });