+++ source/renderer/app/stores/ada/AdaWalletsStore.js
@@ -5,6 +5,7 @@ import Wallet from '../../domain/Wallet';
 import { matchRoute, buildRoute } from '../../utils/routing';
 import Request from '.././lib/LocalizedRequest';
 import { ROUTES } from '../../routes-config';
+import WalletAddDialog from '../../components/wallet/WalletAddDialog';
 import type { walletExportTypeChoices } from '../../types/walletExportTypes';
 import type { WalletImportFromFileParams } from '../../actions/ada/wallets-actions';
 import type { ImportWalletFromFileResponse } from '../../api/ada/index';
@@ -114,7 +115,7 @@ export default class AdaWalletsStore extends WalletStore {
     // ...or keep it open in case it has errored out (so that error message can be shown)
     setTimeout(() => {
       if (!this.restoreRequest.isExecuting) this._setIsRestoreActive(false);
-      if (!this.restoreRequest.isError) this.actions.dialogs.closeActiveDialog.trigger();
+      if (!this.restoreRequest.isError) this._toggleAddWalletDialogOnActiveRestoreOrImport();
     }, this.WAIT_FOR_SERVER_ERROR_TIME);
 
     const restoredWallet = await this.restoreRequest.execute(params).promise;
@@ -139,7 +140,7 @@ export default class AdaWalletsStore extends WalletStore {
     // ...or keep it open in case it has errored out (so that error message can be shown)
     setTimeout(() => {
       if (!this.importFromFileRequest.isExecuting) this._setIsImportActive(false);
-      if (!this.importFromFileRequest.isError) this.actions.dialogs.closeActiveDialog.trigger();
+      if (!this.importFromFileRequest.isError) this._toggleAddWalletDialogOnActiveRestoreOrImport();
     }, this.WAIT_FOR_SERVER_ERROR_TIME);
 
     const { filePath, walletName, walletPassword } = params;
@@ -182,6 +183,19 @@ export default class AdaWalletsStore extends WalletStore {
   }) => {
     if (this.walletExportType !== params.walletExportType) {
       this.walletExportType = params.walletExportType;
+    }
+  };
+
+  // =================== PRIVATE API ==================== //
+
+  _toggleAddWalletDialogOnActiveRestoreOrImport = () => {
+    // Once restore/import is under way we need to either:
+    // A) show the 'Add wallet' dialog (in case we don't have any wallets) or
+    // B) just close the active dialog and unblock the UI
+    if (this.hasLoadedWallets && !this.hasAnyWallets) {
+      this.actions.dialogs.open.trigger({ dialog: WalletAddDialog });
+    } else {
+      this.actions.dialogs.closeActiveDialog.trigger();
     }
   };